--- .vscode/keybindings.json
@@ -1,7 +1,0 @@
-// Place your key bindings in this file to override the defaults
-[
-        { "key": "f5", "command": "workbench.action.reloadWindow" },
-        { "key": "ctrl+shift+c", "command": "workbench.action.tasks.build" },
-        { "key": "ctrl+shift+x", "command": "workbench.action.tasks.runTask", "args": "Compile-Project-X" },
-        { "key": "f1", "command": "workbench.action.tasks.runTask", "args": "launch_offline_help" }
-]
