--- tsProjects/prjNeuroPredict1/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "editor.quickSuggestions": false
-}
