--- tsProjects/prjNeuroPredict1/PythonLib.code-workspace
@@ -1,26 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "../.."
-		},
-		{
-			"path": "../../../MQL5/Brokers/ICMarkets"
-		}
-	],
-	"settings": {
-		"git.ignoreLimitWarning": true,
-		"editor.quickSuggestions": {
-			"comments": "off",
-			"strings": "off",
-			"other": "off"
-		},
-		"files.associations": {
-			"*.mqh": "cpp",
-			"*.mq4": "cpp",
-			"*.mq5": "cpp",
-			"format": "cpp",
-			"filesystem": "cpp",
-			"xstring": "cpp"
-		}
-	}
-}
